+++ src/Types/IDEntityCadastralNumber.php
@@ -17,7 +17,7 @@ class IDEntityCadastralNumber extends AbstractTypedIDEntity implements HasCadast
      */
     final public static function make(string $value, ?string $type = null): self
     {
-        return new static($value);
+        return new static($value, true);
     }
 
     /**