+++ tests/Types/IDEntityChassisTest.php
@@ -7,10 +7,9 @@ namespace AvtoDev\IDEntity\Tests\Types;
 use Illuminate\Support\Str;
 use AvtoDev\IDEntity\IDEntityInterface;
 use AvtoDev\IDEntity\Types\IDEntityChassis;
+use PHPUnit\Framework\Attributes\CoversClass;
 
-/**
- * @covers \AvtoDev\IDEntity\Types\IDEntityChassis
- */
+#[CoversClass(IDEntityChassis::class)]
 class IDEntityChassisTest extends AbstractIDEntityTestCase
 {
     /**
@@ -25,14 +24,23 @@ class IDEntityChassisTest extends AbstractIDEntityTestCase
     {
         $entity = $this->entityFactory();
 
-        $this->assertSame($valid = 'LA130-0128818', $entity::normalize(Str::lower($valid)));
-        $this->assertSame($valid, $entity::normalize(" {$valid}  "));
+        $this->assertSame($valid = 'LA1300128818', $entity::normalize(Str::lower($valid)));
+        $this->assertSame($valid, $entity::normalize(" $valid  "));
         $this->assertSame($valid, $entity::normalize('LA130–0128818'));
         $this->assertSame($valid, $entity::normalize('Lа130-0128818'));
         $this->assertSame($valid, $entity::normalize('LA130--0128818'));
         $this->assertSame($valid, $entity::normalize('LA130-0128№;:?№?*№%$@$%@#818'));
+        $this->assertSame($valid, $entity::normalize(' LA130  0128818'));
 
-        $this->assertSame('LA130 0128818', $entity::normalize(' LA130  0128818'));
+        $this->assertSame('АБВГДЕЁЖЗИЙКЛМНОПРСТУФХЦЧШЩЪЫЬЭЮЯ', $entity::normalize('абвгдеёжзийклмнопрстуфхцчшщъыьэюя'));
+        $this->assertSame('A12BC45', $entity::normalize('  А12-ВС.45  '));
+        $this->assertSame('A123BC45', $entity::normalize('  А123ВС45  '));
+        $this->assertSame('A12BC45', $entity::normalize('А12/В_С-45'));
+        $this->assertSame('ΛΩШЖԱԲ', $entity::normalize('ΛΩШЖԱԲ'));
+        $this->assertSame('УУУШ111', $entity::normalize('YYYШ111'));
+        $this->assertSame('YYYT111', $entity::normalize('УУУТ111'));
+        $this->assertSame('УУУШ111', $entity::normalize('  YYY-Ш.111  '));
+        $this->assertSame('АВСDЕFGНIJКLМNОРQRSТUVWХУZШ', $entity::normalize('ABCDEFGHIJKLMNOPQRSTUVWXYZШ'));
     }
 
     /**
@@ -67,7 +75,6 @@ class IDEntityChassisTest extends AbstractIDEntityTestCase
         return [
             'TSMEYB21S00610448',
             '38:49:924785:832907',
-
             '',
             Str::random(32),
         ];